Migrate tests to new linker error protocol
