+++ src/cmd/build.rs
@@ -434,57 +434,3 @@ mod tests_ci_only {
         })
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::cmd::build::validate_import_section;
-
-    #[test]
-    fn validate_must_catch_panic_import() {
-        // given
-        let contract = r#"
-            (module
-                (type (;0;) (func (param i32 i32 i32)))
-                (import "env" "_ZN4core9panicking5panic17h00e3acdd8048cb7cE" (func (;5;) (type 0)))
-                (func (;5;) (type 0))
-            )"#;
-        let wasm = wabt::wat2wasm(contract).expect("invalid wabt");
-        let module = parity_wasm::deserialize_buffer(&wasm).expect("deserializing must work");
-
-        // when
-        let res = validate_import_section(&module);
-
-        // then
-        assert!(res.is_err());
-        assert!(res
-            .unwrap_err()
-            .to_string()
-            .contains("An unexpected panic function import was found in the contract Wasm."));
-    }
-
-    #[test]
-    fn validate_must_catch_ink_enforce_error_markers() {
-        // given
-        let contract = r#"
-            (module
-                (type (;0;) (func))
-                (import "env" "__ink_enforce_error_for_message_0xAD931DAA" (func $__ink_enforce_error_for_message_0xAD931DAA (type 0)))
-                (import "env" "__ink_enforce_error_for_constructor_0xAD931DCC" (func $__ink_enforce_error_for_message_0xAD931DCC (type 0)))
-            )"#;
-        let wasm = wabt::wat2wasm(contract).expect("invalid wabt");
-        let module = parity_wasm::deserialize_buffer(&wasm).expect("deserializing must work");
-
-        // when
-        let res = validate_import_section(&module);
-
-        // then
-        assert!(res.is_err());
-        let err = res.unwrap_err().to_string();
-        assert!(err.contains(
-            "The ink! message with the selector `0xAD931DAA` contains an invalid trait call."
-        ));
-        assert!(err.contains(
-            "The ink! constructor with the selector `0xAD931DCC` contains an invalid trait call."
-        ));
-    }
-}
